Back/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-debug-toolbar==4.4.6
django-seed==0.3.1
Faker==30.6.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.2
django-bootstrap5~=24.3
django-bootstrap-icons~=0.8.6
django-environ~=0.11.2
djangorestframework~=3.14.0
django-oauth-toolkit~=3.0.1
djangorestframework-simplejwt>=5.4.0
django-cors-headers==4.2.0
dj-rest-auth==5.0.1
django-allauth==0.61.1
python-dotenv==1.0.1
Pillow
gunicorn
whitenoise
reportlab
coverage